refactor data grid sub components
